+++ core/project.scala
@@ -2,8 +2,7 @@
 //> using options "-java-output-version:11", "-Ysafe-init", "-Xmax-inlines:64"
 //> using options "-Werror", "-Wunused:all", "-deprecation", "-feature", -language:noAutoTupling
 
-//> using dep "org.virtuslab::besom-json:0.2.0"
-
+//> using dep "org.virtuslab::besom-json:0.2.2"
 //> using dep "com.lihaoyi::sourcecode:0.3.1"
 //> using dep "com.google.protobuf:protobuf-java-util:3.24.4"
 //> using dep "io.grpc:grpc-netty:1.57.2"
@@ -12,12 +11,11 @@
 //> using dep "com.thesamet.scalapb::scalapb-runtime:0.11.14"
 //> using dep "com.thesamet.scalapb::scalapb-runtime-grpc:0.11.14"
 //> using dep "com.google.guava:guava:32.1.2-jre"
-
 //> using dep "com.outr::scribe:3.11.3"
 //> using dep "com.outr::scribe-file:3.11.3"
 //> using dep "com.lihaoyi::pprint:0.6.6"
-
 //> using test.dep "org.scalameta::munit:1.0.0-M10"
+
 
 //> using publish.name "besom-core"
 //> using publish.organization "org.virtuslab"