+++ core/src/test/scala/besom/internal/EncoderTest.scala
@@ -7,7 +7,6 @@ import besom.types.{EnumCompanion, Label, StringEnum, Output as _, *}
 import besom.util.*
 import com.google.protobuf.struct.*
 import com.google.protobuf.struct.Struct.toJavaProto
-import com.google.protobuf.struct.Value.Kind
 
 object EncoderTest:
 
@@ -26,6 +25,9 @@ object EncoderTest:
 
   case class PlainCaseClass(data: String, moreData: Int) derives Encoder
   case class OptionCaseClass(data: Option[String], moreData: Option[Int]) derives Encoder
+  case class InputOptionalCaseClass(value: Output[Option[String]], data: Output[Option[Map[String, Output[String]]]])
+      derives Encoder,
+        ArgsEncoder
   case class TestArgs(a: Output[String], b: Output[PlainCaseClass]) derives ArgsEncoder
   case class TestOptionArgs(a: Output[Option[String]], b: Output[Option[PlainCaseClass]]) derives ArgsEncoder
 
@@ -69,6 +71,37 @@ class EncoderTest extends munit.FunSuite with ValueAssertions:
     val (_, encoded) = e.encode(None).unsafeRunSync()
 
     assertEqualsValue(encoded, Null)
+  }
+
+  test("encode secret null") {
+    given Context = DummyContext().unsafeRunSync()
+    val e         = summon[Encoder[Output[Option[String]]]]
+
+    val (_, encoded) = e.encode(Output.secret(None)).unsafeRunSync()
+    val expected     = None.asValue.asSecret
+
+    assertEqualsValue(encoded, expected)
+  }
+
+  test("encode secret input map") {
+    given Context = DummyContext().unsafeRunSync()
+    val e         = summon[Encoder[Output[Option[Map[String, String]]]]]
+
+    val data: besom.types.Input.Optional[Map[String, besom.types.Input[String]]] = None
+    val (_, encoded) = e.encode(data.asOptionOutput(isSecret = true)).unsafeRunSync()
+    val expected     = None.asValue.asSecret
+
+    assertEqualsValue(encoded, expected)
+  }
+
+  test("encode class with secret inputs") {
+    given Context = DummyContext().unsafeRunSync()
+    val e         = summon[Encoder[InputOptionalCaseClass]]
+
+    val (_, encoded) = e.encode(InputOptionalCaseClass(Output.secret(None), Output.secret(None))).unsafeRunSync()
+    val expected     = Map().asValue
+
+    assertEqualsValue(encoded, expected)
   }
 
   test("encode special case class with unmangling") {
@@ -164,6 +197,43 @@ class ArgsEncoderTest extends munit.FunSuite with ValueAssertions:
 
     assert(res.isEmpty, res)
     assert(value.fields.isEmpty, value)
+  }
+
+  test("empty secret args") {
+    given Context = DummyContext().unsafeRunSync()
+    val ae        = summon[ArgsEncoder[InputOptionalCaseClass]]
+
+    val (res, value) = ae
+      .encode(
+        InputOptionalCaseClass(Output.secret(None), Output.secret(None)),
+        _ => false
+      )
+      .unsafeRunSync()
+
+    assert(res.isEmpty, res)
+    assert(value.fields.isEmpty, value)
+  }
+
+  test("secrets args") {
+    given Context = DummyContext().unsafeRunSync()
+    val ae        = summon[ArgsEncoder[InputOptionalCaseClass]]
+
+    val (res, value) = ae
+      .encode(
+        InputOptionalCaseClass(
+          Output.secret(Some("secret")),
+          Output.secret(Some(Map("key" -> Output.secret("value"))))
+        ),
+        _ => false
+      )
+      .unsafeRunSync()
+    val expected = Map(
+      "value" -> "secret".asValue.asSecret,
+      "data" -> Map("key" -> "value".asValue.asSecret).asValue.asSecret
+    ).asValue
+
+    assert(res.keySet == Set("value", "data"), res)
+    assertEqualsValue(value.asValue, expected)
   }
 end ArgsEncoderTest
 
@@ -272,11 +342,26 @@ class ProviderArgsEncoderTest extends munit.FunSuite with ValueAssertions:
     given Context = DummyContext().unsafeRunSync()
     val pae       = summon[ProviderArgsEncoder[ProviderArgs]]
 
-    val expected = Map("kubeconfig" -> "abcd".asValue).asValue
-
     val args         = ProviderArgs(kubeconfig = "abcd")
     val (_, encoded) = pae.encode(args, _ => false).unsafeRunSync()
-    assertEqualsValue(Value(Kind.StructValue(encoded)), expected)
+
+    val expected = Map("kubeconfig" -> "abcd".asValue).asValue
+
+    assertEqualsValue(encoded.asValue, expected)
+  }
+
+  // FIXME: This test is failing because the `ProviderArgsEncoder` is not encoding the `kubeconfig` as a secret
+  test("encode secret ProviderArgs".ignore) {
+    given Context = DummyContext().unsafeRunSync()
+    val pae       = summon[ProviderArgsEncoder[ProviderArgs]]
+
+    val args         = ProviderArgs(kubeconfig = Output.secret("abcd"))
+    val (_, encoded) = pae.encode(args, _ => false).unsafeRunSync()
+
+    val expected = Map("kubeconfig" -> "abcd".asValue.asSecret).asValue
+
+    assertNoDiff(pprint(encoded.asValue).plainText, pprint(expected).plainText)
+    assertEqualsValue(encoded.asValue, expected)
   }
 
   test("json ProviderArgsEncoder test") {
@@ -311,4 +396,61 @@ class ProviderArgsEncoderTest extends munit.FunSuite with ValueAssertions:
       """{"b":"true","helper":"{\"int\":1.0}","str":"x"}"""
     )
   }
+
+  test("json secret ProviderArgsEncoder test") {
+    given Context = DummyContext().unsafeRunSync()
+
+    val res = PropertiesSerializer
+      .serializeResourceProperties(
+        ExampleResourceArgs(
+          Output.secret(None),
+          Output.secret(None),
+          Output.secret(None)
+        )
+      )
+      .unsafeRunSync()
+
+    val json = com.google.protobuf.util.JsonFormat
+      .printer()
+      .omittingInsignificantWhitespace()
+      .sortingMapKeys()
+      .print(toJavaProto(res.serialized))
+
+    assertNoDiff(
+      json,
+      """{}"""
+    )
+  }
 end ProviderArgsEncoderTest
+
+class Regression383Test extends munit.FunSuite with ValueAssertions:
+  import Regression383Test.*
+  import ProtobufUtil.*
+
+  test("#383 regression") {
+    given Context = DummyContext().unsafeRunSync()
+    val e         = summon[ArgsEncoder[SecretArgs]]
+
+    val (_, encoded) = e.encode(SecretArgs(), _ => false).unsafeRunSync()
+    val expected     = Map().asValue
+
+    assertEqualsValue(encoded.asValue, expected)
+  }
+
+object Regression383Test:
+  final case class SecretArgs private (
+    data: besom.types.Output[Option[Map[String, String]]]
+  )
+
+  object SecretArgs:
+    def apply(
+      data: besom.types.Input.Optional[Map[String, Input[String]]] = None
+    )(using besom.types.Context): SecretArgs =
+      new SecretArgs(
+        data = data.asOptionOutput(isSecret = true)
+      )
+
+  given encoder(using besom.types.Context): besom.types.Encoder[SecretArgs] =
+    besom.internal.Encoder.derived[SecretArgs]
+  given argsEncoder(using besom.types.Context): besom.types.ArgsEncoder[SecretArgs] =
+    besom.internal.ArgsEncoder.derived[SecretArgs]