+++ examples/kubernetes-guestbook/components/project.scala
@@ -1,0 +1,6 @@
+//> using scala "3.3.1"
+//> using options -Werror -Wunused:all -Wvalue-discard -Wnonunit-statement
+//> using plugin "org.virtuslab::besom-compiler-plugin:0.2.2"
+//> using dep "org.virtuslab::besom-core:0.2.2"
+//> using dep "org.virtuslab::besom-kubernetes:4.8.0-core.0.2"
+//> using exclude "simple/*"