+++ experimental/project.scala
@@ -1,5 +1,6 @@
 //> using scala "3.3.0"
-//> using plugin "org.virtuslab::besom-compiler-plugin:0.2.0"
-//> using dep "org.virtuslab::besom-core:0.2.0"
-//> using dep "org.virtuslab::besom-kubernetes:4.7.1-core.0.2"
+//> using plugin "org.virtuslab::besom-compiler-plugin:0.2.2"
+//> using dep "org.virtuslab::besom-core:0.2.2"
+//> using dep "org.virtuslab::besom-kubernetes:4.8.0-core.0.2"
 //> using dep "io.github.iltotore::iron:2.4.0"
+