+++ integration-tests/resources/config-example/project.scala
@@ -1,4 +1,5 @@
 //> using scala 3.3.1
 //> using options -java-output-version:11 -Werror -Wunused:all -Wvalue-discard -Wnonunit-statement
-//> using plugin org.virtuslab::besom-compiler-plugin:0.2.0
-//> using dep org.virtuslab::besom-core:0.2.0
+//> using plugin org.virtuslab::besom-compiler-plugin:0.2.2
+//> using dep org.virtuslab::besom-core:0.2.2
+