+++ website/src/remark/codeblockVersion.js
@@ -1,8 +1,8 @@
 const visit = require('unist-util-visit');
 
-const codeblockVersion = () => async (ast) => {
+const codeblockVersion = (besomVersion) => () => async (ast) => {
   visit(ast, 'code', (node) => {
-    node.value = node.value.replace('$version', '0.1.0') // TODO read from env: process.env.BESOM_VERSION
+    node.value = node.value.replace('$version', besomVersion)
   })
 }